+++ platzky/plugin/content_transformer.py
@@ -2,13 +2,43 @@
 
 from __future__ import annotations
 
+import re
 from abc import ABC
 
 import jinja2.ext
 
 from platzky.content_types import ContentType
 from platzky.plugin.plugin import PluginBase
-from platzky.shortcodes import Shortcode, make_shortcode_applier
+from platzky.shortcodes import Shortcode, ShortcodeAttrs
+
+_MAX_ATTR_NAME_LEN = 100
+_MAX_ATTR_VALUE_LEN = 2048
+_ATTR_RE = re.compile(rf'([\w-]{{1,{_MAX_ATTR_NAME_LEN}}})="([^"]{{0,{_MAX_ATTR_VALUE_LEN}}})"')
+
+
+def _apply_shortcodes(content: str, shortcodes: dict[str, Shortcode]) -> str:
+    if not shortcodes:
+        return content
+
+    tag_names = "|".join(re.escape(n) for n in shortcodes)
+    pattern = re.compile(
+        rf"\[({tag_names})((?:\s+[\w-]+=\"[^\"]*\")*)\s*\](?:(.*?)\[/\1\])?",
+        re.DOTALL,
+    )
+
+    def _apply(text: str) -> str:
+        def _replace(m: re.Match[str]) -> str:
+            sc = shortcodes[m.group(1)]
+            attrs = ShortcodeAttrs(list(sc.attributes))
+            attrs._values = dict(_ATTR_RE.findall(m.group(2) or ""))
+            inner = m.group(3) or ""
+            if inner:
+                inner = _apply(inner)
+            return sc.render(attrs, inner)
+
+        return pattern.sub(_replace, text)
+
+    return _apply(content)
 
 
 class ContentTransformerPluginBase(PluginBase, ABC):
@@ -37,7 +67,7 @@ class ContentTransformerPluginBase(PluginBase, ABC):
         Returns:
             Transformed content string.
         """
-        return make_shortcode_applier(self.get_supported_shortcodes())(content)
+        return _apply_shortcodes(content, self.get_supported_shortcodes())
 
     def get_supported_shortcodes(self) -> dict[str, Shortcode]:
         """Return shortcode tags this plugin handles.
