+++ platzky/shortcodes/__init__.py
@@ -4,12 +4,10 @@ from platzky.shortcodes.shortcode import (
     Shortcode,
     ShortcodeAttr,
     ShortcodeAttrs,
-    make_shortcode_applier,
 )
 
 __all__ = [
     "Shortcode",
     "ShortcodeAttr",
     "ShortcodeAttrs",
-    "make_shortcode_applier",
 ]
