+++ tests/unit_tests/plugin/test_capability_plugins.py
@@ -19,7 +19,8 @@ from platzky.platzky import create_app_from_config, create_engine
 from platzky.plugin.content_transformer import ContentTransformerPluginBase
 from platzky.plugin.notifier import AttachmentNotifierPluginBase, NotifierPluginBase
 from platzky.plugin.plugin import PluginBase
-from platzky.shortcodes import Shortcode, ShortcodeAttrs, make_shortcode_applier
+from platzky.plugin.content_transformer import _apply_shortcodes
+from platzky.shortcodes import Shortcode, ShortcodeAttrs
 
 # ---------------------------------------------------------------------------
 # Fixtures and helpers
@@ -293,7 +294,7 @@ class TestContentTransformerPluginBase:
                 return {"btag": _BTagSC()}
 
         combined = {**AFilter({}).get_supported_shortcodes(), **BFilter({}).get_supported_shortcodes()}
-        result = make_shortcode_applier(combined)("[atag]x[/atag] [btag]y[/btag]")
+        result = _apply_shortcodes("[atag]x[/atag] [btag]y[/btag]", combined)
         assert result == "A(x) B(y)"
