+++ tests/unit_tests/test_shortcode.py
@@ -1,6 +1,7 @@
 """Tests for the shortcode parser."""
 
-from platzky.shortcodes import Shortcode, ShortcodeAttrs, make_shortcode_applier
+from platzky.plugin.content_transformer import _apply_shortcodes
+from platzky.shortcodes import Shortcode, ShortcodeAttrs
 
 
 def _sc(tag: str) -> Shortcode:
@@ -18,20 +19,20 @@ def _sc(tag: str) -> Shortcode:
 
 class TestApplyShortcodes:
     def test_empty_handlers_returns_content_unchanged(self) -> None:
-        assert make_shortcode_applier({})("hello [foo]bar[/foo]") == "hello [foo]bar[/foo]"
+        assert _apply_shortcodes("hello [foo]bar[/foo]", {}) == "hello [foo]bar[/foo]"
 
     def test_content_without_tags_returned_unchanged(self) -> None:
         sc = _sc("foo")
-        assert make_shortcode_applier({"foo": sc})("<p>Hello world</p>") == "<p>Hello world</p>"
+        assert _apply_shortcodes("<p>Hello world</p>", {"foo": sc}) == "<p>Hello world</p>"
 
     def test_unknown_tag_passes_through(self) -> None:
         sc = _sc("known")
-        result = make_shortcode_applier({"known": sc})("[unknown]text[/unknown]")
+        result = _apply_shortcodes("[unknown]text[/unknown]", {"known": sc})
         assert result == "[unknown]text[/unknown]"
 
     def test_block_tag_content_passed_to_handler(self) -> None:
         sc = _sc("greet")
-        result = make_shortcode_applier({"greet": sc})("[greet]hello[/greet]")
+        result = _apply_shortcodes("[greet]hello[/greet]", {"greet": sc})
         assert result == "[RENDERED:greet:hello]"
 
     def test_void_tag_calls_handler_with_empty_content(self) -> None:
@@ -45,7 +46,7 @@ class TestApplyShortcodes:
                 calls.append((attrs, content))
                 return "<img>"
 
-        make_shortcode_applier({"img": _ImgSC()})('[img url="x.jpg"]')
+        _apply_shortcodes('[img url="x.jpg"]', {"img": _ImgSC()})
         assert calls == [({"url": "x.jpg"}, "")]
 
     def test_attrs_parsed_into_dict(self) -> None:
@@ -59,23 +60,23 @@ class TestApplyShortcodes:
                 received.append(attrs)
                 return ""
 
-        make_shortcode_applier({"foo": _FooSC()})('[foo color="#f00" size="large"]x[/foo]')
+        _apply_shortcodes('[foo color="#f00" size="large"]x[/foo]', {"foo": _FooSC()})
         assert received == [{"color": "#f00", "size": "large"}]
 
     def test_multiple_different_tags_both_replaced(self) -> None:
         a = _sc("a")
         b = _sc("b")
-        result = make_shortcode_applier({"a": a, "b": b})("[a]X[/a] and [b]Y[/b]")
+        result = _apply_shortcodes("[a]X[/a] and [b]Y[/b]", {"a": a, "b": b})
         assert "[RENDERED:a:X]" in result
         assert "[RENDERED:b:Y]" in result
 
     def test_unregistered_tag_between_registered_left_unchanged(self) -> None:
         sc = _sc("foo")
-        result = make_shortcode_applier({"foo": sc})("[unknown]z[/unknown] [foo]x[/foo]")
+        result = _apply_shortcodes("[unknown]z[/unknown] [foo]x[/foo]", {"foo": sc})
         assert "[unknown]z[/unknown]" in result
         assert "[RENDERED:foo:x]" in result
 
     def test_multiline_content_preserved(self) -> None:
         sc = _sc("block")
-        result = make_shortcode_applier({"block": sc})("[block]line1\nline2[/block]")
+        result = _apply_shortcodes("[block]line1\nline2[/block]", {"block": sc})
         assert result == "[RENDERED:block:line1\nline2]"
